move data functions to data.c
