+++ ext/mri/arrayfire.c
@@ -692,105 +692,7 @@ static VALUE arf_get_platform(VALUE self){
   return Qnil;
 }
 
-//Data
-
-
-static VALUE arf_constant(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_constant_complex(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_constant_long(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_constant_ulong(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_range(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_iota(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_identity(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_diag_create(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_diag_extract(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_join(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_join_many(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_tile(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_reorder(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_shift(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_moddims(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_flat(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_flip(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_lower(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_upper(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_select(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_select_scalar_r(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_select_scalar_l(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_replace(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_replace_scalar(VALUE self){
-  return Qnil;
-}
-
+#include "cmodules/data.c"
 #include "cmodules/lapack.c"
 
 // Random
